Refactor account page to avoid long file lengths
